requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
celery==4.4.1
Django>=2.2
djangorestframework==3.14.0
psycopg2>=2.7.5,<2.8.0
python-dotenv==0.21.0
pytz==2022.7.1
redis==3.4.1
sqlparse==0.4.3
tzdata==2022.7
pillow==9.4.0
importlib-metadata==4.13.0